Solve bug retrocompatibility issue on menu hidding
On injector.config.permissions.routes[modelName].menu is instantiated
with an empty array, so it was never entering into the undedefined case
(if). So if a permission for a given model was not entered in the
permissions file, the model was configured as hideMenu=true and
backoffice was empty.

This bug broke compatibility for all RouteInjector projects that didn't
have configured the entry on permissions.